+++ src/routes/v1/apply/accreditation.test.js
@@ -1,8 +1,13 @@
+import {
+  LOGGING_EVENT_ACTIONS,
+  LOGGING_EVENT_CATEGORIES
+} from '../../../common/enums/event.js'
+
 const mockLoggerInfo = vi.fn()
 const mockLoggerError = vi.fn()
 const mockLoggerWarn = vi.fn()
 
-vi.mock('../common/helpers/logging/logger.js', () => ({
+vi.mock('../../../common/helpers/logging/logger.js', () => ({
   createLogger: () => ({
     info: (...args) => mockLoggerInfo(...args),
     error: (...args) => mockLoggerError(...args),
@@ -12,9 +17,9 @@ vi.mock('../common/helpers/logging/logger.js', () => ({
 
 let server
 
-describe('/v1/apply/organisation route', () => {
+describe('/accreditation route', () => {
   beforeAll(async () => {
-    const { createServer } = await import('../server.js')
+    const { createServer } = await import('../../../server.js')
     server = await createServer()
     await server.initialize()
   })
@@ -28,7 +33,7 @@ describe('/v1/apply/organisation route', () => {
 
     const response = await server.inject({
       method: 'POST',
-      url: '/v1/apply/organisation',
+      url: '/v1/apply/accreditation',
       payload
     })
 
@@ -36,13 +41,11 @@ describe('/v1/apply/organisation route', () => {
 
     const body = JSON.parse(response.payload)
     expect(body.success).toBe(true)
-    expect(body.originalPayload).toEqual(payload)
     expect(mockLoggerInfo).toHaveBeenCalledWith({
       message: expect.any(String),
-      payload,
       event: {
-        action: undefined,
-        category: undefined
+        category: LOGGING_EVENT_CATEGORIES.API,
+        action: LOGGING_EVENT_ACTIONS.REQUEST_RECEIVED
       }
     })
   })
@@ -50,7 +53,7 @@ describe('/v1/apply/organisation route', () => {
   it('returns 400 if payload is not an object', async () => {
     const response = await server.inject({
       method: 'POST',
-      url: '/v1/apply/organisation',
+      url: '/v1/apply/accreditation',
       payload: 'not-an-object'
     })
 
@@ -62,7 +65,7 @@ describe('/v1/apply/organisation route', () => {
   it('returns 400 if payload is null', async () => {
     const response = await server.inject({
       method: 'POST',
-      url: '/v1/apply/organisation',
+      url: '/v1/apply/accreditation',
       payload: null
     })
