+++ src/routes/v1/apply/index.js
@@ -1,0 +1,5 @@
+import { organisation } from './organisation.js'
+import { registration } from './registration.js'
+import { accreditation } from './accreditation.js'
+
+export const apply = [organisation, registration, accreditation]
