+++ src/routes/v1/apply/registration.js
@@ -1,16 +1,13 @@
 import Boom from '@hapi/boom'
-import { createLogger } from '../common/helpers/logging/logger.js'
+import { createLogger } from '../../../common/helpers/logging/logger.js'
 import {
   LOGGING_EVENT_ACTIONS,
   LOGGING_EVENT_CATEGORIES
-} from '../common/enums/event.js'
+} from '../../../common/enums/event.js'
 
-/**
- * Test endpoint to receive payloads and respond with status.
- */
-const organisation = {
+const registration = {
   method: 'POST',
-  path: '/v1/apply/organisation',
+  path: '/v1/apply/registration',
   options: {
     validate: {
       payload: (value, _options) => {
@@ -29,16 +26,13 @@ const organisation = {
       event: {
         category: LOGGING_EVENT_CATEGORIES.API,
         action: LOGGING_EVENT_ACTIONS.REQUEST_RECEIVED
-      },
-      payload: request.payload
+      }
     })
 
     return h.response({
-      success: true,
-      receivedAt: new Date().toISOString(),
-      originalPayload: request.payload
+      success: true
     })
   }
 }
 
-export { organisation }
+export { registration }
